refactor(viewer): split CellBuilderController into imperative scene modules
CellBuilderController.tsx was 3,810 lines: a 3,550-line `init()` closure
(29 functions, ~64 getState() calls) wearing a React component. The
component is now 47 lines — it waits for the renderer/scene/camera and
calls initCellBuilderScene(); everything else moved to
components/viewer/cellbuilder/ as plain modules.

Every three.js object and every piece of in-flight interaction state now
lives on ONE explicit scene context (sceneContext.ts), whose type is
inferred from the factory so adding a field needs no second declaration.
Each module takes that context as its first argument instead of closing
over component scope, and reaches the viewer through getViewerRuntime()
and the store exactly as before.

Modules (each header states what it owns): sceneConstants, sceneTypes,
geometryHelpers, sceneContext, builderGrid, overlays, meshRebuild,
cadPreview, portsOverlay, cellGizmo, portGizmo, resizeHandles,
pointerTools, numericEntry, insertTools, keyboard, initScene (the
composer). The exported component name and its (absent) props are
unchanged.

Behaviour-preserving: function bodies are moved verbatim; the edits are
the ctx parameter, the ctx. prefix on shared state, named types for the
four keyboard entry state machines, and stable handler identities in the
composer so add/removeEventListener still pair up. syncOffset now runs
just after the context is built rather than mid-construction — it only
sets the container position, before any frame.

Largest resulting file is 753 lines (keyboard.ts, one key handler).
